DDW-16 Table comparison view for stake pools - fixes for stakepool table, seperating table into header and body
